Refactor global handlers and update project structure
Renamed `GlobalQuadtreeHandler.h` to `GlobalQuadtreeHandler_PG.h` and moved its content. Added `updateQuadtrees` and various getter/setter methods to `GlobalQuadtreeHandler` class. Introduced `GlobalClock` singleton in `GlobalClock_PG.h` and `GlobalContext` struct in `GlobalContext_PG.h`. Added `GlobalTextQueuePipeline` singleton in `GlobalTextQueuePipeline_PG.h`. Updated several files to include `GlobalQuadtreeHandler_PG.h`. Updated `World_PG.h` to use `GlobalContext` and `GlobalTextQueuePipeline`. Modified `LandMassHandler_PG.cpp` to use `GlobalTextQueuePipeline` for loot messages. Updated `RegionHandler_PG.h` and `Region_PG.h` to use `float` literals and `vm::magnitude`. Updated project files to reflect these changes.

+++ PirateGame/Header/GlobalChunkHandler_PG.h
@@ -97,7 +97,7 @@ namespace PirateGame {
 			}
 			// Generate the chunk
 			std::shared_ptr<Chunk> newChunk = std::make_shared<Chunk>(chunkCoord, chunkSize, cellSize);
-			newChunk->setRegionType(regionHandler.generateRegionType(sf::Vector2f(chunkCoord.first, chunkCoord.second)));
+			newChunk->setRegionType(regionHandler.generateRegionType(sf::Vector2f(static_cast<float>(chunkCoord.first), static_cast<float>(chunkCoord.second))));
 			chunks.push_back(newChunk);
 		}
